+++ programs/coinflip/src/instructions/update_oracle.rs
@@ -3,10 +3,10 @@ use anchor_lang::prelude::*;
 pub fn handler(
     ctx: Context<super::UpdateOracle>,
     fee_percentage: u8,
-    oracle_buffer_time: i64,
+    oracle_buffer_time: u16,
     max_players: u16,
-    max_timeout: i64,
-    min_timeout: i64,
+    max_timeout: u16,
+    min_timeout: u16,
 ) -> Result<()> {
     let oracle = &mut ctx.accounts.oracle;
     oracle.fee_percentage = fee_percentage;
